requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
pandas
scipy
pyyaml
tiktoken
python-dotenv
